+++ packages/koop-core/CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* CRD API fom `/datasets/:id`
+* CRD API for `/datasets/:id/metadata`
+
 ## [3.0.0] - 2017-02-13
 ### Added
 * First stable release of Koop 3.0